test: move agent-cleanup-filter test to unit + assert exit code (PER-122)
- Move TC10 (cleanup_unadopted_not_called_for_agent_paths) from
  tests/integration/test_adopt_agent.py to new
  tests/unit/cli/test_adoption_cleanup_filter.py — test is fully
  mocked and sub-second, belongs under unit/
- Add assert result.exit_code == 0 with output + exception in failure
  message so a CLI crash surfaces clearly rather than silently passing
- Fix CHANGELOG.md trailing whitespace (pre-commit)
