+++ libs/beacon/tests/integration/test_adopt_agent.py
@@ -581,83 +581,6 @@ def test_accept_rollback_restores_stale_tool_symlink_target(
 # ---------------------------------------------------------------------------
 
 
-# ---------------------------------------------------------------------------
-# TC10: cleanup_unadopted_artifacts must not receive agents/* paths
-# ---------------------------------------------------------------------------
-
-
-def test_cleanup_unadopted_not_called_for_agent_paths(
-    project: tuple, warehouse: Path, monkeypatch
-):
-    """cleanup_unadopted_artifacts must not receive agents/* paths during adopt.
-
-    Agents are handled atomically inside commit_session() (round-3 atomicity
-    contract). The CLI must filter them out before passing to cleanup.
-    """
-    from unittest.mock import MagicMock
-
-    from beacon.cli.adoption import adopt
-    from beacon.domains.adoption.models import AdoptResult
-    from click.testing import CliRunner
-
-    project_root, artifacts_path, beacon_yaml = project
-
-    beacon_yaml.write_text(
-        "artifacts:\n"
-        "  contexts: []\n"
-        "  skills: []\n"
-        "  agents:\n"
-        "    - agents/spec-planner.md\n"
-    )
-
-    cleanup_received: list[str] = []
-
-    def spy_cleanup(unadoptions, *args, **kwargs):
-        cleanup_received.extend(unadoptions)
-
-    mock_manifest = MagicMock()
-    mock_manifest.artifacts.contexts = []
-    mock_manifest.artifacts.skills = []
-    mock_manifest.artifacts.agents = ["agents/spec-planner.md"]
-
-    mock_result = AdoptResult(
-        to_adopt=[],
-        to_unadopt=["agents/spec-planner.md"],
-        pending_accept=[],
-        pending_reject=[],
-    )
-    mock_app = MagicMock()
-    mock_app.run.return_value = mock_result
-
-    monkeypatch.setattr("beacon.cli.adoption.find_project_root", lambda: project_root)
-    monkeypatch.setattr(
-        "beacon.cli.adoption.WorkspaceConfig",
-        lambda: MagicMock(warehouse=MagicMock(local_path=str(warehouse))),
-    )
-    monkeypatch.setattr(
-        "beacon.cli.adoption.BeaconManifest",
-        MagicMock(from_yaml=MagicMock(return_value=mock_manifest)),
-    )
-    monkeypatch.setattr("beacon.cli.adoption.discover_pending", lambda root: [])
-    monkeypatch.setattr(
-        "beacon.cli.adoption.discover_adoptable", lambda *a, **kw: ([], [])
-    )
-    monkeypatch.setattr("beacon.cli.adoption.is_interactive", lambda: True)
-    monkeypatch.setattr(
-        "beacon.cli.adoption.AdoptApp", lambda *args, **kwargs: mock_app
-    )
-    monkeypatch.setattr("beacon.cli.adoption.commit_session", lambda **kwargs: None)
-    monkeypatch.setattr("beacon.cli.adoption.cleanup_unadopted_artifacts", spy_cleanup)
-
-    runner = CliRunner()
-    runner.invoke(adopt, [])
-
-    agent_paths = [p for p in cleanup_received if p.startswith("agents/")]
-    assert not agent_paths, (
-        f"cleanup_unadopted_artifacts was called with agent paths: {agent_paths}"
-    )
-
-
 def test_adopt_accept_only_agent_updates_gitignore(project: tuple, warehouse: Path):
     """abc adopt accepting an agent must add .gitignore entries even if no skills are involved."""
     project_root, artifacts_path, beacon_yaml = project
